structure: move domain services to server/services/
