+++ server/api/activity.get.ts
@@ -1,5 +1,5 @@
 import { activityQuerySchema } from "#shared/schemas/activity";
-import { getDailyPlaytime } from "~~/lib/activity";
+import { getDailyPlaytime } from "~~/server/services/activity";
 
 export default defineEventHandler(async (event) => {
   const { year } = await getValidatedQuery(event, activityQuerySchema.parse);
