+++ server/api/games/[id]/hidden.patch.ts
@@ -2,7 +2,7 @@ import {
   gameRouterParamsSchema,
   setGameHiddenBodySchema,
 } from "#shared/schemas/games";
-import { setGameHidden } from "~~/lib/games";
+import { setGameHidden } from "~~/server/services/games";
 
 export default defineEventHandler(async (event) => {
   const { id } = await getValidatedRouterParams(
