+++ server/api/games/[id]/state.patch.ts
@@ -2,7 +2,7 @@ import {
   gameRouterParamsSchema,
   setGameStateBodySchema,
 } from "#shared/schemas/games";
-import { setGameState } from "~~/lib/games";
+import { setGameState } from "~~/server/services/games";
 
 export default defineEventHandler(async (event) => {
   const { id } = await getValidatedRouterParams(
