+++ server/api/games/[id]/timeline.get.ts
@@ -1,5 +1,5 @@
 import { gameRouterParamsSchema } from "#shared/schemas/games";
-import { getGameTimeline } from "~~/lib/games";
+import { getGameTimeline } from "~~/server/services/games";
 
 export default defineEventHandler(async (event) => {
   const { id } = await getValidatedRouterParams(
