+++ server/api/games/duplicates.get.ts
@@ -1,4 +1,7 @@
-import { findDuplicatePairs, getDistinctPairs } from "~~/lib/duplicates";
+import {
+  findDuplicatePairs,
+  getDistinctPairs,
+} from "~~/server/services/duplicates";
 
 export default defineEventHandler(async () => {
   return {
