+++ server/api/games/index.get.ts
@@ -1,4 +1,4 @@
-import { getGames } from "~~/lib/games";
+import { getGames } from "~~/server/services/games";
 
 export default defineEventHandler(async () => {
   const games = await getGames();
