+++ server/api/games/split.post.ts
@@ -1,5 +1,5 @@
 import { splitGameBodySchema } from "#shared/schemas/games";
-import { splitGame } from "~~/lib/games";
+import { splitGame } from "~~/server/services/games";
 
 export default defineEventHandler(async (event) => {
   const { provider, providerId } = await readValidatedBody(
