+++ server/api/settings.get.ts
@@ -1,4 +1,4 @@
-import { getSettings } from "~~/lib/settings";
+import { getSettings } from "~~/server/services/settings";
 
 export default defineEventHandler(async () => {
   return await getSettings();
