+++ server/api/settings.patch.ts
@@ -1,5 +1,5 @@
 import { settingsPatchSchema } from "#shared/schemas/settings";
-import { updateSettings } from "~~/lib/settings";
+import { updateSettings } from "~~/server/services/settings";
 
 export default defineEventHandler(async (event) => {
   const patch = await readValidatedBody(event, settingsPatchSchema.parse);
