+++ server/services/games.ts
@@ -2,13 +2,6 @@ import { and, asc, desc, eq, inArray, isNotNull, or } from "drizzle-orm";
 import type { GameState } from "#shared/game-state";
 import { playDayOf } from "#shared/playDay";
 import type { PlaytimeSession } from "#shared/types/PlaytimeSession";
-import { refreshGameAggregates } from "~~/lib/gameAggregates";
-import {
-  deriveSessions,
-  type PlaytimeProviderRow,
-  type PlaytimeSnapshot,
-} from "~~/lib/playtimeTimeline";
-import { getPlayDaySettings } from "~~/lib/settings";
 import { db } from "~~/server/database/client";
 import {
   epicGame,
@@ -23,6 +16,13 @@ import {
   steamGamePlaytime,
 } from "~~/server/database/schema";
 import { countProviderRows } from "~~/server/providers/rows";
+import { refreshGameAggregates } from "~~/server/services/gameAggregates";
+import {
+  deriveSessions,
+  type PlaytimeProviderRow,
+  type PlaytimeSnapshot,
+} from "~~/server/services/playtimeTimeline";
+import { getPlayDaySettings } from "~~/server/services/settings";
 
 export type PlaytimeProvider = "steam" | "gog" | "epic";
